Preparando para um executavel

+++ Manipulando arquivos/BatePonto.py
@@ -6,15 +6,28 @@ import customtkinter
 from PIL import Image, ImageTk
 from tkinter import messagebox
 import mysql.connector
+import os
+import sys
+
+def resource_path(relative_path):
+    """Obtenha o caminho absoluto ao recurso, trabalhando para desenvolvimento e compilado"""
+    try:
+        base_path = sys._MEIPASS
+    except Exception:
+        base_path = os.path.abspath(".")
+    return os.path.join(base_path, relative_path)
 
 def main():
+    
     def cabecalho():
-        with open('Folha de ponto.csv', 'w', newline='') as file:
-            csv.register_dialect('my_dialect', delimiter=';')
-            cabecalho_ = csv.writer(file, dialect='my_dialect')
-            cabecalho_.writerow(["INICIO", "TERMINO", "TEMPO ESTUDADO"])
-    #cabecalho()
-
+        if not os.path.exists('Folha de ponto.csv') or os.path.getsize('Folha de ponto.csv') == 0:
+            with open('Folha de ponto.csv', 'w', newline='') as file:
+                csv.register_dialect('my_dialect', delimiter=';')
+                cabecalho_ = csv.writer(file, dialect='my_dialect')
+                cabecalho_.writerow(["INICIO", "TERMINO", "TEMPO ESTUDADO"])
+    
+    cabecalho()
+    
     def calculando(saida, entrada):
         diferenca = saida - entrada
         return diferenca
@@ -36,12 +49,14 @@ def main():
         def registrando():
             usuario = entrada1.get()
             cd_user = entrada2.get()
-            if not usuario or not cd_user:
-                messagebox.showwarning("FLOOR DIZ:", "Preencha todos os campos.")
+            if not usuario or not cd_user.isdigit():
+                messagebox.showwarning("FLOOR DIZ:", "Preencha todos os campos corretamente. O código de usuário deve ser numérico.")
                 return
-             
+            
+            cd_user = int(cd_user)
+
             conexao = mysql.connector.connect(
-                host='autorack.proxy.rlwy.net',
+                host='autorack.proxy.rlwy.net', 
                 port=35695,
                 user='root',
                 password='HwxRPcthWWGmOcfLZsiyBTpeKlKewPkf', 
@@ -134,7 +149,7 @@ def main():
             entrada2.place_forget()
             bt_voltar.place_forget()
             criando.place_forget()
-
+            recupera_senha.place(x=640, y=610)
         def janela_registro():
             combobox.place_forget()
             entry2.place_forget()
@@ -151,80 +166,136 @@ def main():
             criando.place(x=630, y=420)
             bt_voltar.place(x=630, y=460)
             butao_sair.place(x=650, y=500)
+            recupera_senha.place_forget()
         
         def fechar_janela_principal():
             root.destroy()
 
         def nova_janela(event):  
-            janela2 = tk.Toplevel(root)  # Usar Toplevel em vez de Tk para criar uma nova janela
+            janela2 = tk.Toplevel(root)  
             janela2.grab_set() 
-            image = Image.open("LOGO.png")
-            image = image.resize((80, 80), Image.Resampling.LANCZOS)
-            photo = ImageTk.PhotoImage(image)
-            image_label = tk.Label(janela2, image=photo)
-            image_label.place(x=190, y=15) 
-            def procurando():
-                input_recupera.place_forget()
-                procurar.place_forget()
-                input_criar = customtkinter.CTkEntry(janela2, placeholder_text='CRIE SUA NOVA SENHA:', width=210, height=30)
-                input_criar._set_appearance_mode('light') 
-                input_criar.place(x=130, y=100)
-                guarda = customtkinter.CTkButton(janela2, text='SALVAR SENHA', command=procurando)
-                guarda.place(x=160, y=150)
-            janela2.geometry('500x500')
-            input_recupera = customtkinter.CTkEntry(janela2, placeholder_text='DIGITE SEU USUARIO:', width=210, height=30)
-            input_recupera._set_appearance_mode('light')
-            input_recupera.place(x=130, y=100)
-            procurar = customtkinter.CTkButton(janela2, text='PROCURAR USUARIO', command=procurando)
-            procurar.place(x=160, y=150)
-            janela2.mainloop()  
             
+            
+            def procurando():
+                conexao = mysql.connector.connect(
+                    host='autorack.proxy.rlwy.net',
+                    user='root',
+                    port=35695,
+                    password='HwxRPcthWWGmOcfLZsiyBTpeKlKewPkf',
+                    database='railway',
+                )
+                cursor = conexao.cursor()
+                valor = input_recupera.get()
+                consulta = 'SELECT * FROM tabela_usuarios WHERE nome_user = %s'
+                cursor.execute(consulta, [valor])
+                registro = cursor.fetchone()
+                if registro:
+                    messagebox.showinfo("FLOOR DIZ:", "USUÁRIO ENCONTRADO")
+                    input_recupera.place_forget()
+                    procurar.place_forget()
+                    botao_saida.place_forget()
+                    input_criar = customtkinter.CTkEntry(janela2, placeholder_text='CRIE SUA NOVA SENHA:', width=210, height=30)
+                    input_criar._set_appearance_mode('light') 
+                    input_criar.place(x=130, y=100)
+                    def salvo():
+                        conexao = mysql.connector.connect(
+                      host='autorack.proxy.rlwy.net',
+                      user='root',
+                      port=35695,
+                      password='HwxRPcthWWGmOcfLZsiyBTpeKlKewPkf',
+                      database='railway',
+                      )
+                        cursor = conexao.cursor()
+                        new_pass = input_criar.get()
+                        
+                        if new_pass.isdigit():
+                         try:
+                           comando = 'UPDATE tabela_usuarios SET cd = %s WHERE nome_user = %s'
+                           cursor.execute(comando, [new_pass, valor])
+                           conexao.commit()
+                           messagebox.showinfo("FLOOR DIZ:", "Senha alterada com sucesso")
+                         except mysql.connector.Error as err:
+                          messagebox.showerror("FLOOR DIZ:", f"Erro: {err}")
+                         finally:
+                          cursor.close()
+                          conexao.close()
+                          janela2.destroy()
+                        else:
+                         messagebox.showwarning("FLOOR DIZ:", "A senha deve conter apenas números.")
+
+                    criar = customtkinter.CTkButton(janela2, text='SALVAR', width=110, height=30, command=salvo)
+                    criar.place(x=160, y=150)
+                else:
+                    messagebox.showerror("FLOOR DIZ:", "Usuário não encontrado")
+                cursor.close()
+                conexao.close()
+
+            input_recupera = customtkinter.CTkEntry(janela2, placeholder_text='DIGITE SEU USUÁRIO:', width=210, height=30)
+            input_recupera._set_appearance_mode('light') 
+            input_recupera.place(x=130, y=60)
+
+            procurar = customtkinter.CTkButton(janela2, text='PROCURAR', command=procurando)
+            procurar.place(x=160, y=110)
+
+            botao_saida = customtkinter.CTkButton(janela2, text='SAIR', command=janela2.destroy)
+            botao_saida.place(x=160, y=150)
+            janela2.geometry("500x500")  
+            janela2.title("ALTERAR SENHA")  
+            
+
         root = tk.Tk()
         root.geometry("1994x834")
         root.title("FLOOR")
-        image = Image.open("LOGO.png")
+
+        logo_path = resource_path("resources/LOGO.png")
+        image = Image.open(logo_path)
         image = image.resize((300, 300), Image.Resampling.LANCZOS)
         photo = ImageTk.PhotoImage(image)
         image_label = tk.Label(root, image=photo)
+        image_label.image = photo  # Mantém a referência à imagem
         image_label.place(x=550, y=-20)
-        #ELEMENTOS TELA REGISTROS
-        escolha_entrada = ttk.Label(root, text="TIPO DE ENTRADA: ")
+
+        escolha_entrada = ttk.Label(root, text="TIPO DE ENTRADA::")
         escolha_entrada.place(x=440, y=280)
-        options = ["INICIO", "TÉRMINO"]
-        combobox = ttk.Combobox(root, values=options, state='readonly')
+
+        combobox = ttk.Combobox(root, values=["INICIO", "TERMINO"], state="readonly")
         combobox.place(x=440, y=300, width=500, height=35)
         combobox.set("INICIO")
-        
-        digite_user = ttk.Label(root, text="DIGITE SEU USUÁRIO: ")
+        digite_user = ttk.Label(root, text="DIGITE O SEU USUARIO:")
         digite_user.place(x=440, y=350)
-        
+
         entry2 = ttk.Entry(root)
-        entry2.place(x=440, y=370, width=500, height=35)
-        
-        label_pass = ttk.Label(root, text="DIGITE SEU CÓDIGO DE USUÁRIO: ")
+        entry2.place(x=440, y=370,width=500, height=35)
+
+        label_pass = ttk.Label(root, text="DIGITE SEU CODIGO DE USUÁRIO:")
         label_pass.place(x=440, y=420)
-        
+
         senha = ttk.Entry(root)
-        senha.place(x=440, y=440, width=500, height=35)
-        
-        butao = customtkinter.CTkButton(root, text="REGISTRAR", command=comando)
+        senha.place(x=440, y=440,width=500, height=35)
+
+        butao = customtkinter.CTkButton(root, text="REGISTRAR" ,command=comando)
         butao.place(x=630, y=500)
 
-        butao_sair = customtkinter.CTkButton(root, text='SAIR', command=fechar_janela_principal, width=100)
+        cria_user = customtkinter.CTkButton(root, text="CRIAR USUÁRIO", command=janela_registro)
+        cria_user.place(x=630, y=580)
+
+        butao_sair = customtkinter.CTkButton(root, text="SAIR", command=fechar_janela_principal,width=100)
         butao_sair.place(x=650, y=540)
+         
         recupera_senha = ttk.Label(root, text='ESQUECI A SENHA', font=('Arial', 9, "underline"))
         recupera_senha.place(x=640, y=610)
         recupera_senha.bind("<Button-1>", nova_janela)
-        cria_user = customtkinter.CTkButton(root, text="CRIAR USUÁRIO", command=janela_registro)
-        cria_user.place(x=630, y=580)
-        #ELEMENTOS TELA CRIAR USER
-        entrada1 = customtkinter.CTkEntry(root, placeholder_text='DIGITE SEU NOME:', width=500, height=35)
-        entrada2 = customtkinter.CTkEntry(root, placeholder_text='CRIE SEU CÓDIGO DE USUARIO:', width=500, height=35)
-        bt_voltar = customtkinter.CTkButton(root, text="VOLTAR", command=voltar)
-        criando = customtkinter.CTkButton(root, text='SALVAR', command=registrando)
+        
+        entrada1 = customtkinter.CTkEntry(root, placeholder_text="Digite o nome de usuário:", width=500, height=35)
+        entrada2 = customtkinter.CTkEntry(root, placeholder_text="Crie um código de usuário (Apenas números):", show="*", width=500, height=35)
 
+        criando = customtkinter.CTkButton(root, text="CRIAR",command=registrando)
+        bt_voltar = customtkinter.CTkButton(root, text="VOLTAR", command=voltar)
         root.mainloop()
         
-    create_interface()
+        
 
-main()
+    create_interface()
+    
+main() 
+
